+++ packages/stylist/src/parser.rs
@@ -1,3 +1,4 @@
+use crate::ast::{Block, Rule, RuleContent, ScopeContent, Sheet, StyleAttribute};
 use crate::{Error, Result};
 use nom::{
     branch::alt,
@@ -9,7 +10,6 @@ use nom::{
     sequence::{delimited, preceded, separated_pair, terminated},
     IResult,
 };
-use stylist_core::ast::{Block, Rule, RuleContent, ScopeContent, Sheet, StyleAttribute};
 
 #[cfg(test)]
 use log::trace;